+++ server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -5,9 +5,9 @@ import java.math.BigDecimal;
 public interface AccountDao {
 
     //todo: remove userId from these. they need to pass in what the client inputs (which will be username and switched in the dto)
-    double findBalance(int userId);
+    BigDecimal findBalance(int userId);
 
-    double addToBalance(double amountToAdd, int userId);
+    BigDecimal addToBalance(BigDecimal amountToAdd, int userId);
 
-    double subtractFromBalance(double amountToSubtract, int userId);
+    BigDecimal subtractFromBalance(BigDecimal amountToSubtract, int userId);
 }
